Add structure for a new cytoscape extension
Create a new folder "invariants", where all math code will be. Eventually that can be
published as a cytoscape extension, independent from the UI and this project.
The extension is loaded using cytoscape.use() . We used here the Girth name for the
upcoming feature.

TODO: Make the TS compiler aware of the class extension, currently the compiler complains.