Implement the infobox
The interface "Mode" has been updated to have new functions, see modes.ts.
Now we have a common interface for all modes, despite that they are used
in two places: toolbar and infobox.
There are now two active modes: primaryMode (for the toolbar) and
secondaryMode (for the infobox).
The primaryMode sets the user interaction. The secondaryMode can render
features in the canvas (via the render() method).
All the modes in the infobox have their method infobox() called even if
they are not active, in order to have the infobox updated while the
primaryMethod is acting.